requirements.txt
fastapi
beautifulsoup4
requests
python-dotenv
google-api-python-client